+++ tests/end_to_end_tests/test_database.py
@@ -1,6 +1,7 @@
 import os
 import pandas as pd
 import sqlalchemy
+from sqlalchemy import inspect
 
 from dataregistry import DataRegistry
 from dataregistry.db_basic import SCHEMA_VERSION